refactor(nexus): refactored modules
