+++ crates/nexus/src/domain/aggregate.rs
@@ -1,5 +1,5 @@
-use super::handler::{AggregateCommandHandler, CommandHandlerResponse};
-use crate::core::{Command, DomainEvent, Id};
+use super::{Command, DomainEvent, Id};
+use crate::command::handler::{AggregateCommandHandler, CommandHandlerResponse};
 use smallvec::SmallVec;
 use std::fmt::Debug;
 use thiserror::Error as ThisError;
