+++ crates/nexus/src/infra/event_id.rs
