+++ crates/nexus/src/query/handler.rs
@@ -1,5 +1,5 @@
 use super::repository::ReadModelRepository;
-use crate::core::Query;
+use crate::domain::Query;
 use std::{
     boxed::Box, clone::Clone, error::Error as StdError, marker::PhantomData, pin::Pin,
     task::Context, task::Poll,
